+++ tests/testthat/test-pi-output-mapping.R
@@ -15,37 +15,82 @@ test_that("PIOutputMapping instance prints without errors", {
   expect_no_error(print(outputMapping))
 })
 
-test_that("Read-only fields cannot be set", {
+test_that("PIOutputMapping read-only fields cannot be set", {
   outputMapping <- PIOutputMapping$new(quantity = testQuantity)
   expect_error(outputMapping$observedDataSets <- list(), "is readonly")
   expect_error(outputMapping$dataTransformations <- list(), "is readonly")
   expect_error(outputMapping$quantity <- NULL, "is readonly")
+  expect_error(outputMapping$simId <- NULL, "is readonly")
 })
 
-test_that("Observed data sets can be added", {
-  filePath <- getTestDataFilePath("AciclovirLaskinData.xlsx")
-  dataConfiguration <- ospsuite::createImporterConfigurationForFile(filePath = filePath)
-  dataConfiguration$sheets <- "Laskin 1982.Group A"
-  dataConfiguration$namingPattern <- "{Source}.{Sheet}"
-  observedData <- loadDataSetsFromExcel(
-    xlsFilePath = filePath,
-    importerConfigurationOrPath = dataConfiguration
-  )
-
+test_that("PIOutputMapping adds and removes observed data sets correctly", {
   outputMapping <- PIOutputMapping$new(quantity = testQuantity)
+  observedData <- testObservedData()
+
   expect_no_error(
     outputMapping$addObservedDataSets(
       observedData$`AciclovirLaskinData.Laskin 1982.Group A`
     )
   )
-
   expect_equal(
     outputMapping$observedDataSets[["AciclovirLaskinData.Laskin 1982.Group A"]],
     observedData$`AciclovirLaskinData.Laskin 1982.Group A`
   )
+
+  expect_no_error(outputMapping$removeObservedDataSet(names(observedData)))
+  expect_equal(length(outputMapping$observedDataSets), 0)
 })
 
-test_that("Scaling can be changed to predefined values", {
+test_that("PIOutputMapping adds data without molecular weight and retrieves it", {
+  outputMapping <- PIOutputMapping$new(quantity = testQuantity)
+  observedData <- testObservedData()
+  observedData[[1]]$molWeight <- NA_real_
+
+  expect_no_error(
+    outputMapping$addObservedDataSets(
+      observedData$`AciclovirLaskinData.Laskin 1982.Group A`
+    )
+  )
+  expect_equal(
+    outputMapping$observedDataSets[[1]]$molWeight,
+    225.21
+  )
+})
+
+test_that("PIOutputMapping throws error when unit conversion fails", {
+  outputMapping <- PIOutputMapping$new(quantity = testQuantity)
+  observedData <- testObservedData()
+
+  observedData[[1]]$yDimension <- "Amount"
+  observedData[[1]]$yUnit <- "mol"
+
+  expect_error(
+    outputMapping$addObservedDataSets(
+      observedData$`AciclovirLaskinData.Laskin 1982.Group A`
+    ),
+    "Unit conversion failed for quantity .* and observed data .*"
+  )
+})
+
+test_that("PIOutputMapping errors when mol weight is missing and can't be retrieved", {
+  outputMapping <- PIOutputMapping$new(quantity = testQuantity)
+  observedData <- testObservedData()
+  observedData[[1]]$molWeight <- NA_real_
+
+  mockthat::with_mock(
+    `ospsuite::getMolWeightFor` = function(quantity) stop(),
+    {
+      expect_error(
+        outputMapping$addObservedDataSets(
+          observedData$`AciclovirLaskinData.Laskin 1982.Group A`
+        ),
+        "Unit conversion failed for quantity .* and observed data .*"
+      )
+    }
+  )
+})
+
+test_that("PIOutputMapping allows scaling to be changed to predefined values", {
   outputMapping <- PIOutputMapping$new(quantity = testQuantity)
   outputMapping$scaling <- "log"
   expect_equal(outputMapping$scaling, "log")
@@ -54,36 +99,41 @@ test_that("Scaling can be changed to predefined values", {
   expect_error(outputMapping$scaling <- "invalidScaling", "not in defined")
 })
 
-test_that("Global x-offsets are correctly applied to all datasets", {
+test_that("PIOutputMapping applies global x-offsets to datasets", {
   outputMapping <- PIOutputMapping$new(quantity = testQuantity)
   outputMapping$setDataTransformations(xOffsets = -5)
   expect_equal(outputMapping$dataTransformations$xOffsets, -5)
 })
 
-test_that("Default values for offsets and factors are correctly set", {
+test_that("PIOutputMapping sets default values for offsets and factors correctly", {
   outputMapping <- PIOutputMapping$new(quantity = testQuantity)
+
   expect_equal(outputMapping$dataTransformations$xFactors, 1)
   expect_equal(outputMapping$dataTransformations$yFactors, 1)
   expect_equal(outputMapping$dataTransformations$xOffsets, 0)
   expect_equal(outputMapping$dataTransformations$yOffsets, 0)
 })
 
-test_that("Global x-factors are correctly applied to all datasets", {
+test_that("PIOutputMapping applies global x-factors to datasets", {
   outputMapping <- PIOutputMapping$new(quantity = testQuantity)
   outputMapping$setDataTransformations(xFactors = 2)
+
   expect_equal(outputMapping$dataTransformations$xFactors, 2)
 })
 
-test_that("Global x-offsets and x-factors can be set simultaneously", {
+test_that("PIOutputMapping sets global x-offsets and x-factors simultaneously", {
   outputMapping <- PIOutputMapping$new(quantity = testQuantity)
   outputMapping$setDataTransformations(xFactors = 2, xOffsets = 5)
   expect_equal(outputMapping$dataTransformations$xFactors, 2)
   expect_equal(outputMapping$dataTransformations$xOffsets, 5)
 })
 
-test_that("X-factors can be set in a dataset-specific manner, with labels", {
+test_that("PIOutputMapping sets x-factors in a dataset-specific manner with labels", {
   outputMapping <- PIOutputMapping$new(quantity = testQuantity)
-  outputMapping$setDataTransformations(labels = c("data1", "data2"), xFactors = c(2, 3))
+  outputMapping$setDataTransformations(
+    labels = c("data1", "data2"), xFactors = c(2, 3)
+  )
+
   expect_equal(outputMapping$dataTransformations$xFactors[["data1"]], 2)
   expect_equal(outputMapping$dataTransformations$xFactors[["data2"]], 3)
   expect_equal(outputMapping$dataTransformations$yFactors[["data1"]], 1)
@@ -94,12 +144,12 @@ test_that("X-factors can be set in a dataset-specific manner, with labels", {
   expect_equal(outputMapping$dataTransformations$yOffsets[["data2"]], 0)
 })
 
-test_that("Error is thrown when a non-function is passed to transformResultsFunction", {
+test_that("PIOutputMapping errors when non-function passed to transformResultsFunction", {
   outputMapping <- PIOutputMapping$new(quantity = testQuantity)
   expect_error(outputMapping$transformResultsFunction("invalid"), "non-function")
 })
 
-test_that("transformResultsFunction is correctly applied", {
+test_that("PIOutputMapping applies transformResultsFunction correctly", {
   outputMapping <- PIOutputMapping$new(quantity = testQuantity)
   outputMapping$transformResultsFunction <- function(x) x * 2
   expect_equal(outputMapping$transformResultsFunction(5), 10)